Move RateCheckerParameters to its own file
